Ok V1.3B2

+++ dialWithPHP.ino
@@ -81,7 +81,7 @@ String encodeUri(const String aUri) {
   String answer = "";
   String specialChar = F(".-~_{}[],;:\"\\");
   // uri encode maison :)
-  for (int N = 0; N < aUri.length(); N++) {
+  for (uint16_t N = 0; N < aUri.length(); N++) {
     char aChar = aUri[N];
     //TODO:  should I keep " " to "+" conversion ????  save 2 char but oldy
     if (aChar == ' ') {
